move some logic back to nlpmodel from overrides
